requirements.txt
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
dominate==2.9.1
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools==4.56.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.37.0
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.10.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_311/mkl_random_1676481991689/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
netron==8.0.7
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy==2.2.6
onnx @ file:///D:/bld/onnx_1729538895641/work
opencv-python==4.12.0.88
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
pyparsing==3.2.1
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.14.0
sentry-sdk==2.7.1
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.1
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2025.1
urllib3 @ file:///C:/b/abs_8e4z8_gh1l/croot/urllib3_1715636317140/work
visdom==0.2.4
wandb==0.17.3
websocket-client==1.8.0
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
